+++ app/src/main/res/layout/activity_movie_details.xml
@@ -1,72 +1,77 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:weightSum="1">
 
-    <ImageView
-        android:id="@+id/imageViewDetail"
+    <TableLayout
         android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:layout_weight="0.44"
-        android:scaleType="fitXY"
-        app:srcCompat="@mipmap/ic_launcher" />
+        android:layout_height="match_parent"
+        android:stretchColumns="1">
 
-    <TextView
-        android:id="@+id/textTitle"
-        android:layout_width="match_parent"
-        android:layout_height="61dp"
-        android:layout_marginLeft="2dp"
-        android:layout_marginTop="210dp"
-        android:layout_weight="0.15"
-        android:text="title"
-        android:textSize="30dp" />
+        <ImageView
+            android:id="@+id/imageViewDetail"
+            android:layout_width="match_parent"
+            android:layout_height="200dp"
+            android:layout_weight="0.44"
+            android:elevation="5dp"
+            android:background="@color/colorAccent"
+            android:padding="1dp"
+            android:scaleType="fitXY"
+            app:srcCompat="@mipmap/ic_launcher" />
 
-    <TextView
-        android:id="@+id/textOverview"
-        android:layout_marginTop="280dp"
-        android:layout_width="match_parent"
-        android:layout_height="113dp"
-        android:layout_marginLeft="2dp"
-        android:layout_weight="0.24"
-        android:text="overview" />
+        <TextView
+            android:id="@+id/textTitle"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="2dp"
+            android:layout_marginTop="2dp"
+            android:layout_weight="0.15"
+            android:text="title"
+            android:textSize="30dp" />
+        <TextView
+            android:id="@+id/textOverview"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="2dp"
+            android:text="overview" />
 
-    <TextView
-        android:id="@+id/labelRating"
-        android:layout_marginTop="400dp"
-        android:layout_marginLeft="2dp"
-        android:layout_width="50dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="0.06"
-        android:text="Rating:" />
+        <RelativeLayout android:layout_marginTop="2dp">
+            <TextView
+                android:id="@+id/labelRating"
+                android:layout_marginLeft="2dp"
+                android:layout_width="50dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.06"
+                android:text="Rating:" />
 
-    <TextView
-        android:id="@+id/textRating"
-        android:layout_marginTop="400dp"
-        android:layout_width="94dp"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="60dp"
-        android:layout_weight="0.06"
-        android:text="rating" />
+            <TextView
+                android:id="@+id/textRating"
+                android:layout_width="94dp"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="60dp"
+                android:layout_weight="0.06"
+                android:text="rating" />
+        </RelativeLayout>
 
-    <TextView
-        android:id="@+id/labelReleaseDate"
-        android:layout_marginTop="430dp"
-        android:layout_width="93dp"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="2dp"
-        android:layout_weight="0.07"
-        android:text="Release Date:" />
+        <RelativeLayout android:layout_marginTop="2dp">
+            <TextView
+                android:id="@+id/labelReleaseDate"
+                android:layout_width="93dp"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="2dp"
+                android:layout_weight="0.07"
+                android:text="Release Date:" />
 
-    <TextView
-        android:id="@+id/textReleaseDate"
-        android:layout_marginTop="430dp"
-        android:layout_width="93dp"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="100dp"
-        android:layout_weight="0.07"
-        android:text="release date" />
+            <TextView
+                android:id="@+id/textReleaseDate"
+                android:layout_width="93dp"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="100dp"
+                android:layout_weight="0.07"
+                android:text="release date" />
+            </RelativeLayout>
 
-
-</FrameLayout>
+    </TableLayout>
+</ScrollView>
